2-语料预处理Agent/9-OCR识别算子/dotsocr_inference_only-master/dotsocr_inference_only-master/requirements.txt
# Dots.OCR 推理客户端依赖
# 用于连接已部署的vLLM服务进行文档解析
# Python版本要求: >=3.10

# 核心依赖
# pillow-simd==9.5.0.post2 # 图像处理
pillow
PyMuPDF==1.26.3          # PDF处理
openai==1.90.0           # vLLM API客户端
tqdm==4.67.1             # 进度条
opencv-python==4.12.0.88 # 计算机视觉处理
numpy==2.1.2             # 数值计算
requests==2.32.4         # HTTP请求
json_repair==0.51.0      # JSON格式修复
pyzbar==0.1.9            # 条形码和二维码识别
ImageHash==4.3.2        # 图像哈希计算
aiofiles

# 异步处理
aiohttp==3.12.15         # 异步HTTP客户端
httpx==0.28.1            # HTTP客户端（用于OpenAI）
idna                     # 国际化域名支持 (requests的依赖)